feature/code-style-parser
### What's done:
  * Removed generation to different file
